+++ arches_lingo/src/arches_lingo/pages/HomePage.vue
