+++ arches_lingo/src/arches_lingo/pages/LoginPage.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts">
-import LoginForm from "@/arches_lingo/pages/login/components/LoginForm.vue";
-import LoginLinks from "@/arches_lingo/pages/login/components/LoginLinks.vue";
+import LoginForm from "@/arches_lingo/components/login/LoginForm.vue";
+import LoginLinks from "@/arches_lingo/components/login/LoginLinks.vue";
 </script>
 
 <template>
